v0.2 - Controlled Commenting Behavior
In this version of the reddit bot, commenting behavior is controlled. Before, the bot would comment to any comment in a thread with the matching keyword multiple times. This behavior has been eliminated by storing the comment.id's the bot has already replied to in a file and reading this file before responding to comments again. Other changes include the correction of spelling errors and comments added to the source code for better readability.