Reddit Personal Assistant Bot - v1.0
This is the first iteration of my 'personal assistant' bot. When the command '!save - #tag' is made by me on Reddit, the bot will add the title of the submission to a markdown file on GitHub and sort the posts by the provided '#tag' - making it easier for me to browse my saved posts.